Renamed tracker.component to tracker-character.component
