+++ src/app/app-routing/app-routing.module.ts
@@ -2,7 +2,7 @@ import { AppMasterPageComponent } from '../app-master-page/app-master-page.compo
 import { CharacterListComponent } from '../character-list/character-list.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes, CanActivate } from '@angular/router';
-import { TrackerComponent } from '../tracker/tracker.component';
+import { TrackerCharacterComponent } from '../tracker-character/tracker-character.component';
 import { UserProfileComponent } from '../user-profile/user-profile.component';
 import { AuthGuard } from '../core/auth.guard';
 import { RegisterComponent } from '../register/register.component';
@@ -13,7 +13,7 @@ import { GiveTrackerComponent } from '../give-tracker/give-tracker.component';
 // fill in with { path: 'pathname', component: ComponentName } when adding routes.
 const routes: Routes = [
   { path: 'app', component: AppMasterPageComponent, canActivate: [AuthGuard] },
-  { path: 'tracker', component: TrackerComponent, canActivate: [AuthGuard] },
+  { path: 'tracker', component: TrackerCharacterComponent, canActivate: [AuthGuard] },
   { path: 'login', component: UserProfileComponent},
   { path: 'register', component: RegisterComponent},
   { path: 'accountRecover', component: AccountRecoveryComponent },
